requirements.txt
dj-database-url
Django
django-cors-headers
django-debug-toolbar
django-dynamic-raw-id
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
gunicorn
pillow
psycopg2
pycparser
PyJWT
python-dotenv
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
whitenoise
